requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
boto3==1.18.0
botocore==1.21.0
python-dotenv==0.19.0
Werkzeug==2.0.1
WTForms==2.3.3
pytest==7.0.0
pytest-mock==3.7.0